+++ mm/kfence_core.c
@@ -3,43 +3,31 @@
 #include <asm/pgtable.h>
 #include <asm/tlbflush.h>
 
+#include <linux/list.h>
 #include <linux/debugfs.h>
 #include <linux/moduleparam.h>
 #include <linux/random.h>
 #include <linux/seq_file.h>
+#include <linux/slab.h>
 #include <linux/spinlock_types.h>
 #include <linux/stackdepot.h>
 #include <linux/timer.h>
 
+#include "kfence_core.h"
 #include "slab.h"
 
 /* Usually on, unless explicitly disabled. */
-static bool kfence_enabled;
-
-/*
- * TODO: need to return a freelist back to the cache if it hasn't been used for
- * a while, otherwise we may quickly run out of pages.
- */
-#define MAX_STORED_FREELISTS 256
-struct stored_freelist {
-	struct kmem_cache *cache;
-	void *freelist;
-	int cpu;
-};
-
-#define KFENCE_MAX_CACHES 256
-/*
- * Currently there is less than 100 caches in the running kernel that we need
- * to track. Caches are stored in an array, so that a random cache can be
- * quickly picked.
- */
-static struct kmem_cache *kfence_registered_caches[KFENCE_MAX_CACHES];
-static int kfence_num_caches;
+bool kfence_enabled;
 
 enum kfence_object_state {
 	KFENCE_OBJECT_UNUSED,
 	KFENCE_OBJECT_ALLOCATED,
 	KFENCE_OBJECT_FREED
+};
+
+struct kfence_freelist {
+	struct list_head list;
+	void *obj;
 };
 
 struct alloc_metadata {
@@ -49,11 +37,6 @@ struct alloc_metadata {
 	int size;
 	enum kfence_object_state state;
 };
-
-/* Protects stolen freelists */
-static DEFINE_SPINLOCK(kfence_caches_lock);
-static struct stored_freelist stored_freelists[MAX_STORED_FREELISTS];
-static int num_stored_freelists;
 
 /*
  * It's handy (but not strictly required) that 255 objects with redzones occupy
@@ -67,10 +50,6 @@ static unsigned long kfence_pool_start, kfence_pool_end;
 /* Protects kfence_freelist, kfence_recycle, kfence_metadata */
 static DEFINE_SPINLOCK(kfence_alloc_lock);
 
-struct kfence_freelist {
-	struct list_head list;
-	void *obj;
-};
 /*
  * kfence_freelist is a wrapper around kfence page pointers that allows
  * chaining them.
@@ -97,14 +76,6 @@ static struct alloc_metadata *kfence_metadata;
 /* Order of pages to be allocated when dumping a single object. */
 #define KFENCE_DUMP_ORDER 2
 
-static unsigned long kfence_sample_rate = KFENCE_DEFAULT_SAMPLE_RATE;
-
-#ifdef MODULE_PARAM_PREFIX
-#undef MODULE_PARAM_PREFIX
-#endif
-#define MODULE_PARAM_PREFIX "kfence."
-module_param_named(sample_rate, kfence_sample_rate, ulong, 0444);
-
 /* TODO: there's a similar function in KASAN already. */
 static inline depot_stack_handle_t save_stack(gfp_t flags)
 {
@@ -120,19 +91,11 @@ static inline depot_stack_handle_t save_stack(gfp_t flags)
 	return stack_depot_save(entries, nr_entries, flags);
 }
 
-static noinline void kfence_disable(void)
+noinline void kfence_disable(void)
 {
 	pr_err("Disabling KFENCE\n");
 	WRITE_ONCE(kfence_enabled, false);
 }
-
-#define KFENCE_WARN_ON(cond)                                                   \
-	({                                                                     \
-		bool __cond = WARN_ON(cond);                                   \
-		if (unlikely(__cond))                                          \
-			kfence_disable();                                      \
-		__cond;                                                        \
-	})
 
 /* TODO(glider): kernel_physical_mapping_change() is x86-only */
 unsigned long kernel_physical_mapping_change(unsigned long start,
@@ -194,7 +157,7 @@ static inline bool kfence_unprotect(unsigned long addr)
 	return kfence_change_page_prot(addr, false);
 }
 
-static bool __meminit allocate_pool(void)
+bool __meminit kfence_allocate_pool(void)
 {
 	struct page *pages = NULL;
 	struct kfence_freelist *objects = NULL;
@@ -368,66 +331,6 @@ void guarded_free(void *addr)
 	pr_debug("freed object #%d\n", index);
 }
 
-static struct stored_freelist *find_freelist(struct kmem_cache *c)
-{
-	int i;
-
-	for (i = 0; i < MAX_STORED_FREELISTS; i++) {
-		if (stored_freelists[i].cache == c)
-			return &stored_freelists[i];
-	}
-	return NULL;
-}
-
-static int find_cache(struct kmem_cache *c)
-{
-	int i;
-
-	for (i = 0; i < kfence_num_caches; i++) {
-		if (kfence_registered_caches[i] == c)
-			return i;
-	}
-	return -1;
-}
-
-/* Requires kfence_caches_lock. */
-bool kfence_fix_freelist(struct kmem_cache *s)
-{
-	struct kmem_cache_cpu *c;
-	struct stored_freelist *fl;
-	void *freelist;
-
-	fl = find_freelist(s);
-	if (fl == NULL)
-		return false;
-	freelist = fl->freelist;
-	c = per_cpu_ptr(s->cpu_slab, fl->cpu);
-	/* Nobody else is writing to c->freelist at this point. */
-	WRITE_ONCE(c->freelist, freelist);
-	fl->cache = NULL;
-	num_stored_freelists--;
-	return true;
-}
-
-void *kfence_alloc_and_fix_freelist(struct kmem_cache *s, gfp_t gfp)
-{
-	unsigned long flags;
-	void *ret = NULL;
-
-	if (!READ_ONCE(kfence_enabled))
-		return NULL;
-	spin_lock_irqsave(&kfence_caches_lock, flags);
-	if (!kfence_fix_freelist(s))
-		goto leave;
-	ret = guarded_alloc(s, gfp);
-	spin_unlock_irqrestore(&kfence_caches_lock, flags);
-	pr_debug("kfence_alloc_and_fix_freelist returns %px\n", ret);
-	return ret;
-leave:
-	spin_unlock_irqrestore(&kfence_caches_lock, flags);
-	return NULL;
-}
-
 bool kfence_free(struct kmem_cache *s, struct page *page, void *head,
 		 void *tail, int cnt, unsigned long addr)
 {
@@ -592,65 +495,6 @@ bool kfence_handle_page_fault(unsigned long addr)
 	return kfence_unprotect(addr);
 }
 
-void kfence_cache_register(struct kmem_cache *s)
-{
-	unsigned long flags;
-	int index;
-	const char *name;
-#ifdef CONFIG_MEMCG
-	struct kmem_cache *root;
-#endif
-
-	if (!s)
-		return;
-
-	if (!s->name)
-		name = "ANON";
-	else
-		name = s->name;
-
-#ifdef CONFIG_MEMCG
-	/*
-	 * There are too many memcg child caches, tracking them would require
-	 * too many resources and would reduce the probability of stealing a
-	 * freelist from an "interesting" cache.
-	 * See kfence_observe_memcg_cache() for details about memcg cache
-	 * handling.
-	 */
-	root = s->memcg_params.root_cache;
-	if (root) {
-		pr_debug("skipping memcg cache %s\n", s->name);
-		return;
-	}
-#endif
-
-	if (s->size > PAGE_SIZE) {
-		pr_debug("skipping cache %s because of size: %d\n", name,
-			 s->size);
-		return;
-	}
-	if (s->ctor) {
-		pr_debug("skipping cache %s because of ctor\n", name);
-		return;
-	}
-	if (s->flags & SLAB_TYPESAFE_BY_RCU) {
-		pr_debug("skipping cache %s because of RCU\n", name);
-		return;
-	}
-	spin_lock_irqsave(&kfence_caches_lock, flags);
-	if (kfence_num_caches == KFENCE_MAX_CACHES)
-		goto leave;
-	index = find_cache(s);
-	if (index == -1) {
-		kfence_registered_caches[kfence_num_caches] = s;
-		kfence_num_caches++;
-	}
-	pr_debug("registered cache %s as #%d\n", name, kfence_num_caches - 1);
-leave:
-	spin_unlock_irqrestore(&kfence_caches_lock, flags);
-}
-EXPORT_SYMBOL(kfence_cache_register);
-
 bool kfence_discard_slab(struct kmem_cache *s, struct page *page)
 {
 	if (!is_kfence_addr((unsigned long)page_address(page)))
@@ -658,146 +502,6 @@ bool kfence_discard_slab(struct kmem_cache *s, struct page *page)
 	/* Nothing here for now, but maybe we need to free the objects. */
 	return true;
 }
-
-void kfence_cache_unregister(struct kmem_cache *s)
-{
-	unsigned long flags;
-	int index;
-
-	if (!s)
-		return;
-
-	pr_debug("unregistering cache %s\n", s->name);
-	spin_lock_irqsave(&kfence_caches_lock, flags);
-	if (kfence_num_caches == 0)
-		goto leave;
-	index = find_cache(s);
-	if (index == -1)
-		goto leave;
-	kfence_fix_freelist(s);
-	if (index != kfence_num_caches - 1)
-		kfence_registered_caches[index] =
-			kfence_registered_caches[kfence_num_caches - 1];
-	kfence_num_caches--;
-leave:
-	spin_unlock_irqrestore(&kfence_caches_lock, flags);
-}
-EXPORT_SYMBOL(kfence_cache_unregister);
-
-static struct kmem_cache *kfence_pick_cache(void)
-{
-	int index, wrap = kfence_num_caches;
-	struct kmem_cache *cache;
-
-	if (!kfence_num_caches)
-		return NULL;
-	index = prandom_u32_max(kfence_num_caches);
-	do {
-		cache = kfence_registered_caches[index];
-		index = (index + 1) % kfence_num_caches;
-		wrap--;
-	} while (!cache && wrap);
-
-	return cache;
-}
-
-/* Requires kfence_caches_lock. */
-static void kfence_steal_freelist(struct kmem_cache *cache, int cpu)
-{
-	struct stored_freelist *fl;
-	struct kmem_cache_cpu *c;
-
-	if (num_stored_freelists == MAX_STORED_FREELISTS)
-		return;
-	if (find_freelist(cache))
-		return;
-	fl = find_freelist(NULL);
-	c = per_cpu_ptr(cache->cpu_slab, cpu);
-	if (KFENCE_WARN_ON(!c))
-		return;
-	num_stored_freelists++;
-	fl->cache = cache;
-	fl->cpu = cpu;
-	/*
-	 * We need to atomically read the old value from c->freelist and write
-	 * NULL to it, but SLUB may allocate from this CPU cache and replace
-	 * c->freelist in the meantime. Use cmpxchg loop to ensure fl->freelist
-	 * contains the latest value.
-	 */
-	do {
-		fl->freelist = READ_ONCE(c->freelist);
-	} while (cmpxchg(&c->freelist, fl->freelist, NULL) != fl->freelist);
-	pr_debug("stole freelist from cache %s on CPU%d!\n", cache->name, cpu);
-}
-
-#ifdef CONFIG_MEMCG
-/*
- * Tracking all caches created by memory cgroups may be hard, as there are lots
- * of them. Instead, we only track root (non-memcg) caches. If an allocation is
- * done from a memcg cache, we check if we stole the freelist from its root. In
- * that case we restore the freelist of the root cache and steal the freelist of
- * the memcg cache.
- *
- * TODO: this function is called for every memcg allocation, need to speed it
- * up. Memcg allocations aren't popular in the kernel though.
- * */
-void kfence_observe_memcg_cache(struct kmem_cache *memcg_cache)
-{
-	unsigned long flags;
-	struct kmem_cache *root;
-	char *name;
-	int cpu = raw_smp_processor_id();
-
-	if (!memcg_cache)
-		return;
-
-	if (memcg_cache->name)
-		name = (char *)memcg_cache->name;
-	else
-		name = "ANON";
-
-	root = memcg_cache->memcg_params.root_cache;
-
-	if (!root)
-		/* This is not a valid memcg child cache. */
-		return;
-
-	spin_lock_irqsave(&kfence_caches_lock, flags);
-	if (kfence_fix_freelist(root)) {
-		kfence_steal_freelist(memcg_cache, cpu);
-		pr_debug("stole freelist from memcg cache %s on CPU%d\n",
-			 memcg_cache->name, cpu);
-	}
-	spin_unlock_irqrestore(&kfence_caches_lock, flags);
-}
-EXPORT_SYMBOL(kfence_observe_memcg_cache);
-#endif
-
-static void steal_random_freelist(void)
-{
-	struct kmem_cache *cache;
-	int cpu;
-	unsigned long flags;
-
-	spin_lock_irqsave(&kfence_caches_lock, flags);
-	cache = kfence_pick_cache();
-	cpu = prandom_u32_max(total_cpus);
-	if (!cache)
-		goto leave;
-	kfence_steal_freelist(cache, cpu);
-leave:
-	spin_unlock_irqrestore(&kfence_caches_lock, flags);
-}
-
-static void kfence_heartbeat(struct timer_list *timer)
-{
-	if (!READ_ONCE(kfence_enabled))
-		return;
-
-	steal_random_freelist();
-	mod_timer(timer, jiffies + msecs_to_jiffies(kfence_sample_rate));
-}
-static DEFINE_TIMER(kfence_timer, kfence_heartbeat);
 
 /*
  * debugfs seq_file operations for /sys/kernel/debug/kfence/objects.
@@ -835,10 +539,10 @@ static int obj_show(struct seq_file *seq, void *v)
 		return 0;
 
 	buf = page_address(buf_page);
-	spin_lock_irqsave(&kfence_caches_lock, flags);
+	spin_lock_irqsave(&kfence_alloc_lock, flags);
 	kfence_dump_object(buf, PAGE_SIZE << KFENCE_DUMP_ORDER, index,
 			   &kfence_metadata[index]);
-	spin_unlock_irqrestore(&kfence_caches_lock, flags);
+	spin_unlock_irqrestore(&kfence_alloc_lock, flags);
 	seq_printf(seq, "%s\n", buf);
 
 	__free_pages(buf_page, KFENCE_DUMP_ORDER);
@@ -876,19 +580,12 @@ void __init kfence_create_debugfs(void)
 				   &obj_fops);
 }
 
-void __init kfence_init(void)
-{
-	if (!kfence_sample_rate)
-		/* The tool is disabled. */
-		return;
-
-	if (allocate_pool()) {
-		WRITE_ONCE(kfence_enabled, true);
-		mod_timer(&kfence_timer, jiffies + 1);
-		pr_info("kfence_init done\n");
-	} else {
-		pr_err("kfence_init failed\n");
-	}
-}
-
 device_initcall(kfence_create_debugfs);
+
+unsigned long kfence_sample_rate = KFENCE_DEFAULT_SAMPLE_RATE;
+
+#ifdef MODULE_PARAM_PREFIX
+#undef MODULE_PARAM_PREFIX
+#endif
+#define MODULE_PARAM_PREFIX "kfence."
+module_param_named(sample_rate, kfence_sample_rate, ulong, 0444);
